requirements.txt
Flask
Gunicorn
google.api_core
keras
google.generativeai